Replace MediaPlayer with ExoPlayer.
ExoPlayer has a bug (google/ExoPlayer#91) that requires us to manually
detect the orientation of a video and transform the TextureView.

The fix is encapsulated in VideoOrientationPatcher; that class should
be removed when ExoPlayer gets fixed.

On API 16 and below:
  - Video orientation must be read using a third-party library, e.g.
mp4parser

On API 17 and above:
  - Video orientation can be read using MediaMetadataRetriever

On API 20 and below:
  - ExoPlayer does not rotate the video automatically

On API 21 and above:
  - ExoPlayer rotates the video automatically

On all API levels:
  - Reported video dimensions in onVideoSizeChanged must be swapped if
the video is portrait