Remove #getContext from App.
This method was kept to ease refactoring of the old codebase that used
it quite often. Now it’s no longer required.

When the application context is needed, e.g. for a long-term reference,
the recommended way to get it is Context#getApplicationContext.

For more info about the differences between different Context
instances, see http://possiblemobile.com/2013/06/context/